paper-daily/requirements.txt
requests>=2.31.0
PyYAML>=6.0.0
python-dateutil>=2.8.2
openreview-py>=1.48.0
feedparser>=6.0.10
pytest>=7.0.0
